minor upd

+++ src/main/java/ru/noleg/testnexign/repository/SubscriberRepository.java
@@ -3,6 +3,6 @@ package ru.noleg.testnexign.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.noleg.testnexign.entity.Subscriber;
 
-public interface SubscriberRepository extends JpaRepository<Subscriber, Integer> {
+public interface SubscriberRepository extends JpaRepository<Subscriber, Long> {
     boolean existsByMsisdn(String msisdn);
 }
